refactor: rename package from assistant_service to ai_assistant_service
- Renamed main package directory to ai_assistant_service
- Updated all import statements throughout the codebase
- Updated test directory structure to match new package name
- Fixed configuration files (Dockerfile, Makefile, pyproject.toml)
- Updated documentation references in README.md and CLAUDE.md
- All tests passing with new package structure

This change better reflects the AI-powered nature of the service and
avoids potential naming conflicts with generic "assistant" packages.
