+++ ai_assistant_service/entities/__init__.py
