+++ ai_assistant_service/entities/config.py
