+++ ai_assistant_service/entities/events.py
