+++ ai_assistant_service/entities/headers.py
