+++ ai_assistant_service/entities/message_data.py
