+++ ai_assistant_service/entities/schemas.py
