+++ ai_assistant_service/entities/step_data.py
