+++ ai_assistant_service/repositories/base.py
