+++ ai_assistant_service/server/error_handlers.py
