+++ ai_assistant_service/services/message_parser.py
