+++ ai_assistant_service/services/sse_stream_handler.py
