+++ ai_assistant_service/structured_logging.py
