+++ ai_assistant_service/tools.py
