+++ tests/ai_assistant_service/services/test_message_parser.py
@@ -2,7 +2,7 @@ import pytest
 from openai.types.beta.threads import Message, TextContentBlock
 from openai.types.beta.threads.text import Text
 
-from assistant_service.services.message_parser import MessageParser
+from ai_assistant_service.services.message_parser import MessageParser
 
 
 @pytest.mark.asyncio
